requirements.txt

aiohttp
sqlalchemy
requests
python-dotenv
        